+++ tests/test_ops/test_image/test_transforms.py
@@ -4,8 +4,7 @@ import cv2 as cv
 import numpy as np
 import pytest
 
-from creevey.ops.image import centercrop, resize, rotate_180, rotate_270, rotate_90, trim_padding
-
+from creevey.ops.image import centercrop, flip_horiz, flip_vert, resize, rotate_180, rotate_270, rotate_90, trim_padding
 
 class TestResize:
     def test_resize_validation(self, sample_image_square_rgb):
@@ -130,6 +129,34 @@ class TestTrimPadding:
         np.testing.assert_almost_equal(actual, sample_image_square_rgb)
 
 
+class TestFlipHoriz:
+    def test_flip_horiz_rgba(self, sample_image_square_rgba, sample_image_square_rgba_flipped_horiz):
+        image = flip_horiz(sample_image_square_rgba)
+        np.testing.assert_almost_equal(image, sample_image_square_rgba_flipped_horiz, decimal=-1)
+
+    def test_flip_horiz_rgb(self, sample_image_square_rgb, sample_image_square_rgb_flipped_horiz):
+        image = flip_horiz(sample_image_square_rgb)
+        np.testing.assert_almost_equal(image, sample_image_square_rgb_flipped_horiz, decimal=-1)
+
+    def test_flip_horiz_grayscale(self, sample_image_tall_grayscale, sample_image_grayscale_flipped_horiz):
+        image = flip_horiz(sample_image_tall_grayscale)
+        np.testing.assert_almost_equal(image, sample_image_grayscale_flipped_horiz, decimal=-1)
+
+
+class TestFlipVert:
+    def test_flip_vert_rgba(self, sample_image_square_rgba, sample_image_square_rgba_flipped_vert):
+        image = flip_vert(sample_image_square_rgba)
+        np.testing.assert_almost_equal(image, sample_image_square_rgba_flipped_vert, decimal=-1)
+
+    def test_flip_vert_rgb(self, sample_image_square_rgb, sample_image_square_rgb_flipped_vert):
+        image = flip_vert(sample_image_square_rgb)
+        np.testing.assert_almost_equal(image, sample_image_square_rgb_flipped_vert, decimal=-1)
+
+    def test_flip_vert_grayscale(self, sample_image_tall_grayscale, sample_image_grayscale_flipped_vert):
+        image = flip_vert(sample_image_tall_grayscale)
+        np.testing.assert_almost_equal(image, sample_image_grayscale_flipped_vert, decimal=-1)
+        
+
 class TestRotate_90:
     def test_rotate_rgba_90(self, sample_image_square_rgba, sample_image_square_rgba_rotated_90):
         image = rotate_90(sample_image_square_rgba)
@@ -170,3 +197,4 @@ class TestRotate_270:
     def test_rotate_grayscale_270(self, sample_image_tall_grayscale, sample_image_tall_grayscale_rotated_270):
         image = rotate_270(sample_image_tall_grayscale)
         np.testing.assert_almost_equal(image, sample_image_tall_grayscale_rotated_270, decimal=-2)
+